finish testing backend, user admin and authorization
this was a big one. introducing testing node's inbuilt test module. user
login and token based authentication was added. both notes app and blogs
apps were updated. user model was added.

+++ bloglist-app/backend/controllers/blogPosts.js
@@ -1,14 +1,28 @@
 const blogRouter = require('express').Router();
+const jwt = require('jsonwebtoken');
+
+const middleware = require('../utils/middleware');
 
 const Blog = require('../models/blog');
 
 blogRouter.get('/', async (req, res) => {
-	const blogs = await Blog.find();
+	const blogs = await Blog.find().populate('user', {
+		username: 1,
+		name: 1,
+		id: 1
+	});
 
 	res.status(200).json(blogs);
 });
 
-blogRouter.post('/', async (req, res) => {
+blogRouter.post('/', middleware.userExtractor, async (req, res) => {
+	const token = req.token;
+	const decodedToken = jwt.verify(token, process.env.SECRET);
+	if (!decodedToken.id) {
+		return res.status(401).json({ error: 'token invalid' });
+	}
+	const user = req.user;
+
 	const body = req.body;
 	if (!body.title || !body.url) {
 		return res
@@ -20,15 +34,32 @@ blogRouter.post('/', async (req, res) => {
 		body.likes = 0;
 	}
 
-	let blog = new Blog(body);
+	const nextBlog = new Blog({
+		title: body.title,
+		author: body.author,
+		url: body.url,
+		likes: body.likes,
+		user: user.id
+	});
+
+	let blog = new Blog(nextBlog);
 
 	blog = await blog.save();
 
 	res.status(201).json(blog);
 });
 
-blogRouter.delete('/:id', async (req, res) => {
+blogRouter.delete('/:id', middleware.userExtractor, async (req, res) => {
+	const token = req.token;
 	const id = req.params.id;
+
+	const decodedToken = jwt.verify(token, process.env.SECRET);
+
+	if (!decodedToken.id) {
+		return res.status(401).json({ error: 'token invalid' });
+	}
+
+	const userId = req.user.id;
 
 	const blog = await Blog.findById(id);
 
@@ -36,6 +67,10 @@ blogRouter.delete('/:id', async (req, res) => {
 		return res
 			.status(404)
 			.json({ error: `blog with id: ${id} doesn't exist` });
+	}
+
+	if (blog.user.toString() !== userId.toString()) {
+		return res.status(401).json({ error: 'unauthorized' });
 	}
 
 	await Blog.findByIdAndDelete(id);
@@ -43,7 +78,7 @@ blogRouter.delete('/:id', async (req, res) => {
 	res.status(204).json(blog);
 });
 
-blogRouter.put('/:id', async (req, res) => {
+blogRouter.put('/:id', middleware.userExtractor, async (req, res) => {
 	const id = req.params.id;
 
 	const blog = await Blog.findById(id);
@@ -54,10 +89,14 @@ blogRouter.put('/:id', async (req, res) => {
 			.json({ error: `blog with id: ${id} doesn't exist` });
 	}
 
+	if (blog.user.toString() !== req.user.id.toString()) {
+		return res.status(401).json({ error: 'unauthorized' });
+	}
+
 	blog.likes++;
 
 	await blog.save();
-	res.status(201).json(blog);
+	res.status(200).json(blog);
 });
 
 module.exports = blogRouter;
